Fixes sorting tests.  All tests pass
